refactor episode story and entry updates to new classes
