+++ test/jobs/story_update_job_test.rb
@@ -26,7 +26,7 @@ describe StoryUpdateJob do
     mock_episode = Minitest::Mock.new
     mock_episode.expect(:copy_media, true)
     mock_episode.expect(:podcast, podcast)
-    Episode.stub(:create_from_story!, mock_episode) do
+    EpisodeStoryHandler.stub(:create_from_story!, mock_episode) do
       podcast.stub(:create_publish_task, true) do
         job.receive_story_update(JSON.parse(body))
       end
